requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
mysql-connector-python==8.0.27
python-telegram-bot==20.7